# packages/api-contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Futrob Private API
  version: 0.1.0
  description: Private HTTP API for Futrob clients (web, React Native / Expo,
    scripts). Not a public third-party API.
servers:
  - url: /api/v1
    description: Same Worker as apps/web
security:
  - bearerAuth: []
...
paths:
...
  /organizations/{organizationId}/competitions:
    get:
      operationId: listOrganizationCompetitions
      tags:
        - competitions
      summary: List competitions in an organization
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization competitions newest-first
          content:
            application/json:
              schema:
                type: object
                required:
                  - competitions
                properties:
                  competitions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Competition"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: createCompetitionDraft
      tags:
        - competitions
      summary: Create a competition draft in an organization
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CompetitionDraftInput"
                - type: object
                  properties:
                    creationKey:
                      type: string
                      pattern: ^[A-Za-z0-9_-]{8,120}$
                      description: "Client key: a retry returns the same draft and reuses the cover
                        upload key"
      responses:
        "201":
          description: Competition draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
  /competitions/mine:
    get:
      operationId: listAccessibleCompetitions
      tags:
        - competitions
      summary: List competitions where the actor has a contextual membership
      responses:
        "200":
          description: Only competitions directly accessible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAccessibleCompetitionsResponse"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
      parameters:
        - *a2
  /competitions/explore:
    get:
      operationId: exploreCompetitions
      tags:
        - competitions
      summary: List published competitions discoverable by any authenticated actor
      parameters:
        - *a2
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum:
              - league
              - knockout
              - groups-knockout
              - league-playoffs
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - registration
              - published
              - paused
              - finished
        - name: region
          in: query
          required: false
          schema:
            type: string
            enum:
              - america
              - south-america
              - north-central-america
              - europe
              - africa
              - asia
              - middle-east
              - oceania
        - name: platform
          in: query
          required: false
          schema:
            type: string
            enum:
              - playstation
              - xbox
              - pc
              - nintendo-switch-1
              - nintendo-switch-2
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - updated-desc
              - name-asc
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 48
            default: 24
      responses:
        "200":
          description: Discoverable competitions matching the filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreCompetitionsResponse"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
  /competitions/explore/{competitionId}:
    get:
      operationId: getExploreCompetition
      tags:
        - competitions
      summary: Get one discoverable competition
      parameters:
        - *a2
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Discoverable competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreCompetition"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  /competitions/explore/{competitionId}/application:
    get:
      operationId: getMyCompetitionApplication
      tags:
        - competitions
      summary: Get the actor's own entry in a discoverable competition, if any
      parameters:
        - *a2
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The actor's application or null
          content:
            application/json:
              schema:
                type: object
                required:
                  - application
                properties:
                  application:
                    oneOf:
                      - $ref: "#/components/schemas/CompetitionApplication"
                      - type: "null"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: applyToCompetition
      tags:
        - competitions
      summary: Apply with a new team to a competition in registration; the actor
        becomes its captain and the entry stays pending
      parameters:
        - *a2
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - teamName
                - creationKey
              properties:
                teamName:
                  type: string
                  minLength: 1
                  maxLength: 120
                creationKey:
                  type: string
                  minLength: 8
                  maxLength: 200
      responses:
        "201":
          description: Pending application (idempotent per creationKey)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionApplication"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
...
  /organizations/{organizationId}/competitions/{competitionId}:
    get:
      operationId: getCompetitionDraft
      tags:
        - competitions
      summary: Get an organization-scoped competition draft
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Competition and current rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
    patch:
      operationId: updateCompetitionDraft
      tags:
        - competitions
      summary: Update the editable structure and rules of a competition draft
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompetitionDraftInput"
      responses:
        "200":
          description: Saved draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/teams:
    get:
      operationId: listOrganizationTeams
      tags:
        - teams
      summary: List teams available to the organization operator
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: object
                required:
                  - teams
                properties:
                  teams:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
          headers:
            X-Request-ID: *a1
        "403":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/competitions/{competitionId}/standings:
    get:
      operationId: getCompetitionStandings
      tags:
        - statistics
      summary: Read the official competition standings snapshot
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Standing snapshot or null when nothing has been projected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompetitionStandingsResponse"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/competitions/{competitionId}/team-statistics:
    get:
      operationId: getCompetitionTeamStatistics
      tags:
        - statistics
      summary: Read official team competition statistics
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team competition aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompetitionTeamStatisticsResponse"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/competitions/{competitionId}/rankings:
    get:
      operationId: getCompetitionRankings
      tags:
        - statistics
      summary: Read competition player ranking snapshots
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum:
              - scorer
              - assister
              - rating
              - mvp
              - goalkeeper
      responses:
        "200":
          description: Ranking snapshots for the competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompetitionRankingsResponse"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
...
  /organizations/{organizationId}/competitions/{competitionId}/cover:
    put:
      operationId: updateCompetitionCover
      tags:
        - competitions
      summary: Replace the cover in any status except archived
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cover
              properties:
                cover:
                  $ref: "#/components/schemas/CompetitionCover"
      responses:
        "200":
          description: Competition with its new cover
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional in MVP scaffold; required when identity is wired.
...
  schemas:
...
    CompetitionDraftInput:
      type: object
      required:
        - name
        - gameEdition
        - platform
        - region
        - timeZone
        - format
      properties:
        teams:
          $ref: "#/components/schemas/CompetitionTeamRange"
        schedule:
          $ref: "#/components/schemas/CompetitionSchedule"
        cover:
          $ref: "#/components/schemas/CompetitionCover"
        name:
          type: string
          minLength: 1
          maxLength: 120
        gameEdition:
          type: string
          minLength: 1
          maxLength: 40
        platform:
          type: string
          enum:
            - playstation
            - xbox
            - pc
            - nintendo-switch-1
            - nintendo-switch-2
        region:
          type: string
          enum:
            - america
            - south-america
            - north-central-america
            - europe
            - africa
            - asia
            - middle-east
            - oceania
        timeZone:
          type: string
          minLength: 1
          maxLength: 100
        format:
          type: string
          enum:
            - league
            - knockout
            - groups-knockout
            - league-playoffs
...
    Competition:
      type: object
      required:
        - id
        - organizationId
        - name
        - status
        - modality
        - gameEdition
        - platform
        - region
        - timeZone
        - format
        - teams
        - schedule
        - cover
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - draft
            - registration
            - published
            - paused
            - finished
            - archived
        modality:
          type: string
          const: fc-clubs
        gameEdition:
          type: string
        platform:
          type: string
          enum:
            - playstation
            - xbox
            - pc
            - nintendo-switch-1
            - nintendo-switch-2
        region:
          type: string
          enum:
            - america
            - south-america
            - north-central-america
            - europe
            - africa
            - asia
            - middle-east
            - oceania
        timeZone:
          type: string
        format:
          type: string
          enum:
            - league
            - knockout
            - groups-knockout
            - league-playoffs
        teams:
          $ref: "#/components/schemas/CompetitionTeamRange"
        schedule:
          $ref: "#/components/schemas/CompetitionSchedule"
        cover:
          $ref: "#/components/schemas/CompetitionCover"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListAccessibleCompetitionsResponse:
      type: object
      required:
        - competitions
      properties:
        competitions:
          type: array
          items:
            type: object
            required:
              - competition
              - role
            properties:
              competition:
                $ref: "#/components/schemas/Competition"
              role:
                type: string
                enum:
                  - staff
                  - captain
                  - vice_captain
                  - player
    ExploreCompetition:
      type: object
      required:
        - competition
        - organization
        - approvedTeamCount
      properties:
        competition:
          $ref: "#/components/schemas/Competition"
        organization:
          type: object
          required:
            - id
            - name
          properties:
            id:
              type: string
            name:
              type: string
        approvedTeamCount:
          type: integer
          minimum: 0
    ExploreCompetitionsResponse:
      type: object
      required:
        - items
        - total
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExploreCompetition"
        total:
          type: integer
          minimum: 0
        nextCursor:
          type:
            - string
            - "null"
...
    CompetitionDraft:
      type: object
      required:
        - competition
        - rules
      properties:
        competition:
          $ref: "#/components/schemas/Competition"
        rules:
          $ref: "#/components/schemas/CompetitionRules"
    UpdateCompetitionDraftInput:
      allOf:
        - $ref: "#/components/schemas/CompetitionDraftInput"
        - type: object
          required:
            - rules
          properties:
            rules:
              type: object
              required:
                - regularStage
                - knockoutStage
                - maxRosterSize
              properties:
                regularStage:
                  anyOf:
                    - $ref: "#/components/schemas/CompetitionMatchRules"
                    - type: "null"
                knockoutStage:
                  anyOf:
                    - $ref: "#/components/schemas/CompetitionMatchRules"
                    - type: "null"
                maxRosterSize:
                  anyOf:
                    - type: integer
                      minimum: 1
                    - type: "null"
    Team:
      type: object
      required:
        - id
        - organizationId
        - name
        - createdAt
      properties:
        id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CompetitionApplication:
      type: object
      required:
        - entryId
        - status
        - teamId
        - teamName
        - createdAt
      properties:
        entryId:
          type: string
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
        teamId:
          type: string
        teamName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    GetCompetitionStandingsResponse:
      type: object
      required:
        - standings
      properties:
        standings:
          anyOf:
            - $ref: "#/components/schemas/CompetitionStandingSnapshot"
            - type: "null"
...
    GetCompetitionTeamStatisticsResponse:
      type: object
      required:
        - teams
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/TeamCompetitionStats"
...
    GetCompetitionRankingsResponse:
      type: object
      required:
        - rankings
      properties:
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/RankingSnapshot"
...
  responses:
    ApiError:
      description: API error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...